allow empty string as input for `ParseIdentity`
This commit changes the `ParseIdentity` and `Identity.String`
functions to allow the empty string as valid identity.

`ParseIdentity` returns the Identity zero value for an empty string
and `Identity.String` returns the empty string if the identity is
the zero value.

This simplifies parsing in many cases where an identity string might
be empty.

Signed-off-by: Andreas Auernhammer <github@aead.dev>