add support for boringcrypto FIPS build mode
This commit adds the `internal/boringcrypto` package
that exposes a single constant `FIPS=true` if the package
is built with the build tag `boringcrypto` and `FIPS=false`
otherwise.

The Go standard library (unofficially) supports a FIPS module
via the `GOEXPERIMENT=boringcrypto` env. variable. If set, the
Go compiler includes all source files with the build tag
```
go:build boringcrypto
```

In this "FIPS mode", we don't support Ed25519 since it's a cryptographic
primitive that cannot be used as per the FIPS 140-2 standard. EDDSA over
the curves P-256, P-384 and P-521 is still supported.

This needs to be re-visited once golang/go#69536 lands.

Signed-off-by: Andreas Auernhammer <github@aead.dev>